+++ src/main/java/baseball/domain/ball/BallList.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.ball;
 
 import java.util.ArrayList;
 import java.util.List;
