+++ src/test/java/baseball/domain/BallListTest.java
@@ -2,6 +2,8 @@ package baseball.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import baseball.domain.ball.Ball;
+import baseball.domain.ball.BallList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
